Fix compiler warnings on g++

+++ rice/detail/NativeAttributeSet.hpp
@@ -26,8 +26,8 @@ namespace Rice
       void operator=(const NativeAttribute_T&) = delete;
       void operator=(NativeAttribute_T&&) = delete;
 
-      Resolved matches(int argc, const VALUE* argv, VALUE self) override;
-      VALUE operator()(int argc, const VALUE* argv, VALUE self) override;
+      Resolved matches(size_t argc, const VALUE* argv, VALUE self) override;
+      VALUE operator()(size_t argc, const VALUE* argv, VALUE self) override;
       std::string toString() override;
 
     protected:
